+++ src/Operations.ts
@@ -517,7 +517,8 @@ export class Operations {
       removedCells.push({address, cellType: this.getClipboardCell(address)})
     }
 
-    this.dependencyGraph.removeColumns(columnsToRemove)
+    const {affectedArrays, contentChanges} = this.dependencyGraph.removeColumns(columnsToRemove)
+    this.columnSearch.applyChanges(contentChanges.getChanges())
     this.columnSearch.removeColumns(columnsToRemove)
 
     let version: number
@@ -526,6 +527,8 @@ export class Operations {
       transformation.performEagerTransformations(this.dependencyGraph, this.parser)
       version = this.lazilyTransformingAstService.addTransformation(transformation)
     })
+
+    this.rewriteAffectedMatrices(affectedArrays)
 
     return {
       version: version!,
@@ -584,14 +587,17 @@ export class Operations {
       return
     }
 
-    this.dependencyGraph.addColumns(addedColumns)
+    const {affectedArrays, contentChanges} = this.dependencyGraph.addColumns(addedColumns)
     this.columnSearch.addColumns(addedColumns)
+    this.columnSearch.applyChanges(contentChanges.getChanges())
 
     this.stats.measure(StatType.TRANSFORM_ASTS, () => {
       const transformation = new AddColumnsTransformer(addedColumns)
       transformation.performEagerTransformations(this.dependencyGraph, this.parser)
       this.lazilyTransformingAstService.addTransformation(transformation)
     })
+
+    this.rewriteAffectedMatrices(affectedArrays)
   }
 
   public getOldContent(address: SimpleCellAddress): [SimpleCellAddress, ClipboardCell] {