+++ test/column-index.spec.ts
@@ -746,4 +746,88 @@ describe('Arrays', () => {
       [3, 2]
     ], engine)
   })
+
+  it('should move array values when adding columns before array', () => {
+    const engine = HyperFormula.buildFromArray([
+      ['=-C2:D2'],
+      [null, 'foo', 1, 2]
+    ], {useArrayArithmetic: true, useColumnIndex: true})
+
+    engine.addColumns(0, [0, 1])
+
+    expectColumnIndexToMatchSheet([
+      [null, -1, -2],
+      [null, null, 'foo', 1, 2]
+    ], engine)
+  })
+
+  it('should not move array values when adding columns', () => {
+    const engine = HyperFormula.buildFromArray([
+      ['=-C2:E2', null, null, -3, -2, -1],
+      [null, 'foo', 1, 2, 3]
+    ], {useArrayArithmetic: true, useColumnIndex: true})
+
+    engine.addColumns(0, [1, 1])
+
+    expectColumnIndexToMatchSheet([
+      [-1, -2, -3, null, -3, -2, -1],
+      [null, null, 'foo', 1, 2, 3]
+    ], engine)
+  })
+
+  it('should move array values when removing columns before array', () => {
+    const engine = HyperFormula.buildFromArray([
+      [null, '=-C2:D2'],
+      [null, null, 1, 2]
+    ], {useArrayArithmetic: true, useColumnIndex: true})
+
+    engine.removeColumns(0, [0, 1])
+
+    expectColumnIndexToMatchSheet([
+      [-1, -2],
+      [null, 1, 2]
+    ], engine)
+  })
+
+  it('should not move array values when removing columns', () => {
+    const engine = HyperFormula.buildFromArray([
+      ['=-D2:E2', null, null, -2, -1],
+      [-2, 'foo', null, 1, 2]
+    ], {useArrayArithmetic: true, useColumnIndex: true})
+
+    engine.removeColumns(0, [1, 1])
+
+    expectColumnIndexToMatchSheet([
+      [-1, -2, -2, -1],
+      [-2, null, 1, 2]
+    ], engine)
+  })
+
+  it('should remove array values from index when removing column with left corner', () => {
+    const engine = HyperFormula.buildFromArray([
+      ['=-B2:C2'],
+      [null, 1, 2],
+    ], {useArrayArithmetic: true, useColumnIndex: true})
+
+    engine.removeColumns(0, [0, 1])
+
+    expectColumnIndexToMatchSheet([
+      [],
+      [1, 2],
+    ], engine)
+  })
+
+  it('should remove array values from index when REF after removing columns', () => {
+    const engine = HyperFormula.buildFromArray([
+      ['=-B2:C2', null, 1],
+      [2, null, 3]
+    ], {useArrayArithmetic: true, useColumnIndex: true})
+
+    engine.removeColumns(0, [1, 1])
+
+    expectColumnIndexToMatchSheet([
+      [null, 1],
+      [2, 3]
+    ], engine)
+  })
 })